changed Integer to Long cause speciality field in the database has type int(64)

+++ dao/src/main/java/com/github/KostyaTr/hospital/dao/DoctorSpecialityDao.java
@@ -4,5 +4,5 @@ package com.github.KostyaTr.hospital.dao;
 import java.util.List;
 
 public interface DoctorSpecialityDao {
-    List<Integer> getDoctorsSpecialities(Long doctorId);
+    List<Long> getDoctorsSpecialities(Long doctorId);
 }
